added changes from PR comments
