+++ src/controllers/LoginController.ts
@@ -1,9 +1,8 @@
 import {Container} from "typedi";
-import {Authorized, BadRequestError, Controller, Get, QueryParam, CurrentUser} from "routing-controllers";
+import {Authorized, BadRequestError, Controller, Get, QueryParam} from "routing-controllers";
 import {UserService} from "../services/UserService";
 import {ChallengeRepository} from "../repositories/ChallengeRepository";
 import { LoggerService } from "../services/LoggerService";
-import { User } from "../models/User";
 
 @Controller("/api/login")
 export class LoginController {
