+++ src/views/partials/scripts.ejs
@@ -3,5 +3,4 @@
 <script src="/assets/js/bootstrap.min.js"></script>
 <script src="/assets/js/bootstrap-select.min.js"></script>
 <script src="/assets/js/jquery-ui.min.js"></script>
-<script src="/events/NavUpdater.js"></script>
-<script src="/events/GeneralScriptBundled.js"></script>
+<script src="/events/NavUpdaterBundled.js"></script>
