+++ tools/bundleFrontendModules.ts
@@ -1,7 +1,7 @@
 import * as shell from "shelljs";
 
 // Copy all the view templates
-shell.exec("browserify dist/events/GeneralScript.js > dist/events/GeneralScriptBundled.js");
+shell.exec("browserify dist/events/NavUpdater.js > dist/events/NavUpdaterBundled.js");
 shell.exec("browserify dist/events/LoginScript.js > dist/events/LoginScriptBundled.js");
 shell.exec("browserify dist/events/RegisterScript.js > dist/events/RegisterScriptBundled.js");
 shell.exec("browserify dist/events/AccountScript.js > dist/events/AccountScriptBundled.js");
